+++ OctopusGame/bullet.cpp
@@ -25,8 +25,8 @@ Bullet::Bullet(GraphicalOctopus &graphics, Vector2 spawnPoint, Direction facing,
 }
 
 Bullet::Bullet(GraphicalOctopus &graphics, Player &player, float speed, Force force) :
-	Bullet(graphics, Vector2((int)(player.getX() + player.getBoundingBox().getWidth()/2), 
-	(int)(player.getY() + player.getBoundingBox().getHeight()/2)), player.getFacing(), speed)
+	Bullet(graphics, Vector2((player.getX() + player.getBoundingBox().getWidth()/2), 
+	(player.getY() + player.getBoundingBox().getHeight()/2)), player.getFacing(), speed)
 {
 	this->_player = player;
 	if (this->_facing == RIGHT) {
@@ -39,7 +39,7 @@ Bullet::Bullet(GraphicalOctopus &graphics, Player &player, float speed, Force fo
 }
 
 void Bullet::draw(GraphicalOctopus &graphics) {
-	AnimatedSprite::draw(graphics, (int)(this->_x), (int)(this->_y));
+	AnimatedSprite::draw(graphics, (this->_x), (this->_y));
 }
 
 void Bullet::update(int elapsedTime) {
